# api/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Gatte management API
  version: 1.6.0
  summary: The backend every Gatte front talks to (design/adr/0040).
  description: |
    Served by `mcp-gateway admin` over UNIX sockets only. There is no TCP
    listener. HTTP/1.1, JSON bodies, snake_case fields, times in RFC 3339
    UTC.

    Two sockets, two processes, two directories, two disjoint route sets.
    Every operation below carries `x-gatte-socket`:

    - `operator`: `/run/mcp-gateway-admin/operator/operator.sock`, served
      by a process running as the gateway's service account; the file is
      group `gatte-operators`, `0660`, in a directory of its own.
    - `accounts`: `/run/mcp-gateway-admin/accounts/accounts.sock`, served
      only by a root process; the file is `root:root 0600`, or
      `root:ACCOUNT_GROUP 0660` when `[admin] account_group` delegates
      accounts (both settings are required; the backend refuses a mismatch
      at startup). Its directory is `root:root 0755`.

    A route of the other socket answers 404 `wrong_socket` with
    `details.socket`; a route that does not exist answers 404
    `unknown_route`.

    ## Who serves

    `pkg/adminapi` reads the server's credentials after connecting and
    refuses an accounts socket whose server is not uid 0 (under systemd the
    kernel reports systemd, which created the socket). `WithServerUID` pins
    the operator socket's server.

    ## Who the operator is

...
    `GET /v1/whoami` returns what the backend decided.

...

...
servers:
...
components:
  parameters:
    GatteFront:
      name: Gatte-Front
      in: header
      required: false
      description: Label of the calling front, for the operator row's tag. Display only.
      schema:
        type: string
        pattern: '^[a-z0-9._-]{1,32}$'
...
  responses:
    Error:
      description: |
        Any failure. `code` is stable and machine-readable; `message` is
        one English sentence for a person and may change.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    Overview:
      type: object
      required: [attention, counts, recent_denied]
      properties:
        attention:
          type: array
          description: |
            `serve_not_reporting` first, then changed tools, pending tools,
            unsigned backends, unavailable backends, backends in
            maintenance, and `gateway_maintenance` last.
          items:
            $ref: '#/components/schemas/Attention'
        counts:
          type: object
          required: [tools, approved_usable, review, upstreams, blocked]
          properties:
            tools: {type: integer}
            approved_usable: {type: integer}
            review: {type: integer}
            upstreams: {type: integer}
            blocked: {type: integer}
        recent_denied:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/AuditRecord'
        problems:
          type: array
          description: Parts of the overview that could not be read; the rest is still returned.
          items:
            type: object
            required: [code, part, message]
            properties:
              code:
                type: string
                description: Open set, e.g. `store_busy`, `corrupt_state`, `internal`.
              part:
                type: string
                x-gatte-open-enum: true
                enum: [tools, upstreams, blocks, audit, health, maintenance]
              message:
                type: string
                description: English fallback.
        health:
          description: Present with feature `backend_health`. Absent when it could not be read (see `problems`, part `health`).
          $ref: '#/components/schemas/Health'
...
    ToolList:
      type: object
      required: [tools, counts]
      properties:
        tools:
          type: array
          items: {$ref: '#/components/schemas/Tool'}
        counts:
          type: object
          required: [all, review, approved]
          properties:
            all: {type: integer}
            review: {type: integer}
            approved: {type: integer}
...
    ToolReview:
      type: object
      description: |
        Everything an approval decides on, read in ONE read of the entry.
        `observed.fingerprint` is the value to send to approve.
      required: [tool, observed, callable_by, review_text]
      properties:
        tool: {$ref: '#/components/schemas/Tool'}
        observed: {$ref: '#/components/schemas/Definition'}
        approved:
          description: Present when an approved fingerprint exists and differs from the observed one.
          $ref: '#/components/schemas/Definition'
        diff:
          type: array
          description: |
            approved (`del`) to observed (`add`), changed lines with one line
            of context and `gap` where unchanged lines were left out. A label
            line left as the last context line of a hunk is already dropped.
          items: {$ref: '#/components/schemas/DiffLine'}
        callable_by:
          type: array
          description: |
            Roles that can call this tool once it is approved. Empty means
            approving makes it servable but callable by nobody. For a
            sensitive tool (1.5.0): only the roles marked `non_read` that
            name it in `tools`, which is who it is callable by once CLEARED
            -- a wildcard or a `[role.grants]` name never reaches a
            sensitive tool, so `wildcard` is never true here and
            `wildcard_grant` is never raised for one.
          items: {$ref: '#/components/schemas/RoleCoverage'}
        review_text:
          type: string
          description: |
            The `tool show` text, ALREADY ESCAPED by `internal/visible`, for a
            plain-text details disclosure only. Do not parse it for hidden
            code points (a literal `\u{202E}` and a real one look the same
            here); use the `Text` and `segments` fields.
    ApproveRequest:
      type: object
      required: [server, tool, fingerprint]
      additionalProperties: false
      properties:
        server: {type: string}
        tool: {type: string}
        fingerprint:
          type: string
          description: The fingerprint the operator was shown (ToolReview.observed.fingerprint). Required; the backend never approves "whatever is advertised now".
    ApproveResult:
      allOf:
        - $ref: '#/components/schemas/ActionResult'
        - type: object
          required: [previous_status, tool]
          properties:
            previous_status: {$ref: '#/components/schemas/ToolStatus'}
            previous_baseline: {type: string, description: The approved fingerprint that is no longer accepted (changed tools).}
            tool: {$ref: '#/components/schemas/Tool'}
            callable_by:
              type: array
              items: {$ref: '#/components/schemas/RoleCoverage'}
    ToolReviewSet:
      type: object
      description: |
        Every tool of one backend that waits for review (pending or
        changed), each as `ToolReview` shows it, read in ONE read of the
        backend's entries, in tool-name order (design/adr/0043). `manifest`
        names exactly this set: it is the hex SHA-256 over the backend, the
        number of tools and each tool's name, status, observed fingerprint
        and approved baseline. Show every entry before offering to approve
        it; there is no approval of a set without its manifest.
      required: [server, manifest, tools, pending, changed, hidden_code_points, approvable]
      properties:
        server: {type: string}
        manifest:
          type: string
          description: The value to send to approve-set. Empty when nothing waits.
          pattern: '^([0-9a-f]{64})?$'
        tools:
          type: array
          items: {$ref: '#/components/schemas/ToolReview'}
        pending: {type: integer}
        changed: {type: integer}
        hidden_code_points:
          type: integer
          description: Sum of `observed.hidden_code_points` over `tools`.
        approvable:
          type: boolean
          description: false when nothing waits, or when an observed definition was not kept (`reason` says which).
        reason: {type: string}
    ApproveSetRequest:
      type: object
      required: [server, manifest]
      additionalProperties: false
      properties:
        server: {type: string}
        manifest:
          type: string
          description: The manifest the operator was shown (ToolReviewSet.manifest). Requests may carry the `sha256:` prefix.
...
    ToolRef:
      type: object
      required: [server, tool]
      additionalProperties: false
      properties:
        server: {type: string}
        tool: {type: string}
    RevokeResult:
      allOf:
        - $ref: '#/components/schemas/ActionResult'
        - type: object
          required: [tool]
          properties:
            was_approved_at: {type: string}
            tool: {$ref: '#/components/schemas/Tool'}
...
paths:
...
  /v1/overview:
    get:
      operationId: overview
      tags: [overview]
      x-gatte-socket: [operator]
      summary: What needs attention, the counts, and the latest refusals.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Overview'}
        default: {$ref: '#/components/responses/Error'}
  /v1/tools:
    get:
      operationId: listTools
      tags: [tools]
      x-gatte-socket: [operator]
      summary: The quarantine entries (`tool list`).
      parameters:
        - {name: server, in: query, schema: {type: string}, description: Only this backend.}
        - name: show
          in: query
          schema: {type: string, enum: [all, review, approved], default: all}
          description: '`review` is every entry not approved; `approved` the rest. `counts` is always over all.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolList'}
        default: {$ref: '#/components/responses/Error'}
  /v1/tools/review:
    get:
      operationId: reviewTool
      tags: [tools]
      x-gatte-socket: [operator]
      summary: One tool's definitions, diff and role coverage (`tool show`).
      parameters:
        - {name: server, in: query, required: true, schema: {type: string}}
        - {name: tool, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: OK. When the observed definition was not kept, `observed.kept` is false and approve will answer definition_unavailable.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolReview'}
        default: {$ref: '#/components/responses/Error'}
  /v1/tools/approve:
    post:
      operationId: approveTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Approve the fingerprint that was reviewed, and only that one. Writes `(tool approve)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApproveRequest'}
      responses:
        '200':
          description: |
            Approved, or already approved at exactly this fingerprint (`changed`
            false). A sensitive tool (1.5.0) is approved and still `usable:
            false` until cleared with `clearTool`; the warning
            `sensitive_uncleared` says so. That is not an error.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApproveResult'}
        default:
          description: fingerprint_required, fingerprint_mismatch, fingerprint_moved, definition_unavailable, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/review-set:
    get:
      operationId: reviewToolSet
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Every pending and changed tool of one backend, and the manifest of that set (`tool review -server`). Feature `tool_review_set`.
      parameters:
        - {name: server, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: OK. Nothing waiting is `tools` empty, `manifest` empty and `approvable` false.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ToolReviewSet'}
        default:
          description: invalid_argument (no server), not_found (no tool observed on it), ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/approve-set:
    post:
      operationId: approveToolSet
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: |
        Approve one backend's whole review set, all of it or nothing, and only
        while it is still the set the manifest names. Writes one
        `(tool approve)` row per tool and one `(tool approve set)` row.
        Feature `tool_review_set`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApproveSetRequest'}
      responses:
        '200':
          description: |
            Approved, or nothing waited any more (`changed` false). Approving
            a set never clears (1.5.0): a sensitive tool of it is approved and
            `usable: false` until cleared one at a time with `clearTool`; the
            warning `sensitive_uncleared` names them.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApproveSetResult'}
        default:
          description: manifest_required, manifest_mismatch, definition_unavailable, not_found, corrupt_state, store_busy, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/tools/revoke:
    post:
      operationId: revokeTool
      x-gatte-idempotent: true
      tags: [tools]
      x-gatte-socket: [operator]
      summary: Return an approved tool to pending. Writes `(tool revoke)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ToolRef'}
      responses:
        '200':
          description: Revoked, or already pending (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RevokeResult'}
        default:
          description: changed_not_revocable, not_found, corrupt_state, ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /v1/access/blocks:
    get:
      operationId: listBlocks
      tags: [access]
      x-gatte-socket: [operator]
      summary: The blocklist (`access list`).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [blocks]
                properties:
                  blocks:
                    type: array
                    items: {$ref: '#/components/schemas/Block'}
        default: {$ref: '#/components/responses/Error'}
  /v1/access/block:
    post:
      operationId: blockSubject
      x-gatte-idempotent: true
      tags: [access]
      x-gatte-socket: [operator]
      summary: Block an analyst from the next request on (design/adr/0031). Writes `(access block)`.
      description: The block is placed first and recorded second. If the record fails the answer is still 200 with `recorded` false and warning `audit_write_failed`, and the block stays.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BlockRequest'}
      responses:
        '200':
          description: Blocked, or already blocked (`changed` false). Warning `never_seen` when the trail has no row from this subject.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BlockResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/access/unblock:
    post:
      operationId: unblockSubject
      x-gatte-idempotent: true
      tags: [access]
      x-gatte-socket: [operator]
      summary: Lift a block. Writes `(access unblock)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BlockRequest'}
      responses:
        '200':
          description: Unblocked, or was not blocked (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BlockResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/audit:
    get:
      operationId: listAudit
      tags: [audit]
      x-gatte-socket: [operator]
      summary: The trail, filtered (`audit -json`).
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, default: 200}
          description: Newest N matching records; clamped to 5000.
        - {name: since, in: query, schema: {type: string, format: date-time}}
        - name: subject
          in: query
          schema: {type: string}
          description: Exact match on analyst_identity only, never on analyst_name (design/adr/0037).
        - {name: outcome, in: query, schema: {$ref: '#/components/schemas/AuditOutcome'}}
        - {name: source, in: query, schema: {type: string}, description: Exact source address.}
        - name: until
          in: query
          schema: {type: string, format: date-time}
          description: 1.4.0 (`audit_filters`). Only records BEFORE this time (exclusive; `since` is inclusive). Must be after `since` (`bad_request`).
        - name: tool
          in: query
          schema: {type: string}
          description: 1.4.0 (`audit_filters`). Exact `tool`, e.g. `casemgmt.list_cases` or `(access block)`.
        - name: server
          in: query
          schema: {type: string}
          description: 1.4.0 (`audit_filters`). Exact `target_upstream`. An older backend ignores these three; a front must not send them without the feature.
        - name: before
          in: query
          schema: {type: integer, minimum: 1}
          description: Only records with a `position` lower than this (from `next_before`), to walk the trail backwards.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditPage'}
        default: {$ref: '#/components/responses/Error'}
  /v1/audit/verify:
    post:
      operationId: verifyAudit
      tags: [audit]
      x-gatte-socket: [operator]
      summary: Walk the hash chain and, if given, compare the head (`audit -verify`). Writes nothing.
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VerifyRequest'}
      responses:
        '200':
          description: The verification ran; read `intact` and `head_matches`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VerifyResult'}
        default: {$ref: '#/components/responses/Error'}
...
  /v1/quota:
    get:
      operationId: quotaUsage
      tags: [quota]
      x-gatte-socket: [operator]
      summary: Quota spend per analyst and account (`quota usage`).
      parameters:
        - {name: analyst, in: query, schema: {type: string}}
        - {name: account, in: query, schema: {type: string}}
        - {name: since, in: query, schema: {type: string, format: date-time}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [usage]
                properties:
                  usage:
                    type: array
                    items: {$ref: '#/components/schemas/QuotaUsage'}
        default: {$ref: '#/components/responses/Error'}
...